requirements.txt
crewai
crewai-tools
groq
yfinance
streamlit
newsapi-python
python-dotenv